Refactor simple_embedding_run with separate create_sample_device()
simple_embedding_run with different backend targets shares common
code struction except how the device is created. Refactor the code with
simple_embedding_backend = {simple_embedding, device_backend} so the
common interface/code can be shared.
